Move some domain designers and confirmation dialogs
