+++ src/org/labkey/test/pages/experiment/CreateDataClassPage.java
@@ -18,7 +18,7 @@ package org.labkey.test.pages.experiment;
 import org.labkey.test.WebDriverWrapper;
 import org.labkey.test.WebTestHelper;
 import org.labkey.test.components.domain.DomainFormPanel;
-import org.labkey.test.components.experiment.DataClassDesigner;
+import org.labkey.test.components.ui.domainproperties.dataclasses.DataClassDesigner;
 import org.labkey.test.pages.LabKeyPage;
 import org.labkey.test.params.FieldDefinition;
 import org.openqa.selenium.WebDriver;
