+++ src/org/labkey/test/pages/experiment/CreateSampleTypePage.java
@@ -2,7 +2,7 @@ package org.labkey.test.pages.experiment;
 
 import org.labkey.test.WebDriverWrapper;
 import org.labkey.test.WebTestHelper;
-import org.labkey.test.components.experiment.SampleTypeDesigner;
+import org.labkey.test.components.ui.domainproperties.samples.SampleTypeDesigner;
 import org.openqa.selenium.WebDriver;
 
 public class CreateSampleTypePage extends SampleTypeDesigner<CreateSampleTypePage>
